Major breaking changes
I've lost all the commits and changelogs from this branch, which was on
an old hard drive that failed. :cry:
Needless to say, it's significantly improved.

+++ vCommands/Parsing/Tokens.cs
@@ -9,7 +9,7 @@ namespace vCommands.Parsing
     /// Possible types of tokens in a console command string.
     /// </summary>
     public enum TokenTypes
-        : sbyte
+        : int
     {
         /// <summary>
         /// Separates a sequence of commands.
